Move test functionality from hydra-tx to hydra-tx:testlib
