+++ hydra-node/src/Hydra/API/ClientInput.hs
@@ -6,6 +6,7 @@ import Hydra.Prelude
 import Test.Hydra.Prelude
 
 import Hydra.Tx (ConfirmedSnapshot, IsTx (..), TxIdType)
+import Test.Hydra.Tx.Gen ()
 import Test.QuickCheck.Arbitrary.ADT (ToADTArbitrary)
 
 data ClientInput tx
