+++ hydra-node/src/Hydra/API/ServerOutput.hs
@@ -23,11 +23,12 @@ import Hydra.Tx (HeadId, Party, Snapshot, SnapshotNumber, getSnapshot)
 import Hydra.Tx qualified as Tx
 import Hydra.Tx.ContestationPeriod (ContestationPeriod)
 import Hydra.Tx.Crypto (MultiSignature)
-import Hydra.Tx.IsTx (ArbitraryIsTx, IsTx (..))
+import Hydra.Tx.IsTx (IsTx (..))
 import Hydra.Tx.OnChainId (OnChainId)
 import Hydra.Tx.Snapshot (ConfirmedSnapshot (..), Snapshot (..))
 import Hydra.Tx.Snapshot qualified as HeadState
 import Test.Hydra.Prelude
+import Test.Hydra.Tx.Gen (ArbitraryIsTx)
 import Test.QuickCheck (recursivelyShrink)
 import Test.QuickCheck.Arbitrary.ADT (ToADTArbitrary)
 
@@ -92,7 +93,7 @@ instance IsChainState tx => FromJSON (ClientMessage tx) where
         { omitNothingFields = True
         }
 
-instance (IsChainState tx, ArbitraryIsTx tx) => Arbitrary (ClientMessage tx) where
+instance (IsChainState tx, Arbitrary (ChainStateType tx), ArbitraryIsTx tx) => Arbitrary (ClientMessage tx) where
   arbitrary = genericArbitrary
 
 -- | A friendly welcome message which tells a client something about the
@@ -231,7 +232,7 @@ instance (ArbitraryIsTx tx, Arbitrary (ChainStateType tx)) => Arbitrary (ServerO
   arbitrary = genericArbitrary
   shrink = recursivelyShrink
 
-instance (ArbitraryIsTx tx, IsChainState tx) => ToADTArbitrary (ServerOutput tx)
+instance (ArbitraryIsTx tx, Arbitrary (ChainStateType tx), IsChainState tx) => ToADTArbitrary (ServerOutput tx)
 
 -- | Whether or not to include full UTxO in server outputs.
 data WithUTxO = WithUTxO | WithoutUTxO
