+++ hydra-node/src/Hydra/HeadLogic/Input.hs
@@ -7,9 +7,9 @@ import Test.Hydra.Prelude
 
 import Hydra.API.ClientInput (ClientInput)
 import Hydra.Chain (ChainEvent)
-import Hydra.Chain.ChainState (IsChainState)
+import Hydra.Chain.ChainState (ChainStateType, IsChainState)
 import Hydra.Network.Message (Message, NetworkEvent)
-import Hydra.Tx.IsTx (ArbitraryIsTx)
+import Test.Hydra.Tx.Gen (ArbitraryIsTx)
 
 type TTL = Natural
 
@@ -33,6 +33,6 @@ deriving stock instance IsChainState tx => Show (Input tx)
 deriving anyclass instance IsChainState tx => ToJSON (Input tx)
 deriving anyclass instance IsChainState tx => FromJSON (Input tx)
 
-instance (ArbitraryIsTx tx, IsChainState tx) => Arbitrary (Input tx) where
+instance (ArbitraryIsTx tx, Arbitrary (ChainStateType tx), IsChainState tx) => Arbitrary (Input tx) where
   arbitrary = genericArbitrary
   shrink = genericShrink
