+++ hydra-node/src/Hydra/HeadLogic/State.hs
@@ -17,13 +17,13 @@ import Hydra.Tx (
   Party,
  )
 import Hydra.Tx.Crypto (Signature)
-import Hydra.Tx.IsTx (ArbitraryIsTx)
 import Hydra.Tx.Snapshot (
   ConfirmedSnapshot,
   Snapshot (..),
   SnapshotNumber,
   SnapshotVersion,
  )
+import Test.Hydra.Tx.Gen (ArbitraryIsTx)
 
 -- | The main state of the Hydra protocol state machine. It holds both, the
 -- overall protocol state, but also the off-chain 'CoordinatedHeadState'.
