+++ hydra-node/src/Hydra/Node/State.hs
@@ -12,7 +12,7 @@ import Hydra.Tx (
   HeadId,
   IsTx (..),
  )
-import Hydra.Tx.IsTx (ArbitraryIsTx)
+import Test.Hydra.Tx.Gen (ArbitraryIsTx)
 import Test.QuickCheck (recursivelyShrink)
 
 type PendingDeposits tx = Map (TxIdType tx) (Deposit tx)
