+++ hydra-plutus/test/Hydra/Data/ContestationPeriodSpec.hs
@@ -1,4 +1,5 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
+
 module Hydra.Data.ContestationPeriodSpec where
 
 import Hydra.Prelude
@@ -7,10 +8,10 @@ import Hydra.Data.ContestationPeriod (
   contestationPeriodFromDiffTime,
   contestationPeriodToDiffTime,
  )
+import Hydra.Plutus.Gen ()
 import Test.Hspec (Spec, describe)
 import Test.Hspec.QuickCheck (prop)
 import Test.QuickCheck ((===))
-import Hydra.Plutus.Gen ()
 
 spec :: Spec
 spec = do
