+++ hydra-plutus/testlib/Hydra/Plutus/Gen.hs
@@ -1,12 +1,13 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
+
 module Hydra.Plutus.Gen where
 
 import Hydra.Prelude
 
 import Data.ByteString qualified as BS
-import Hydra.Data.ContestationPeriod (ContestationPeriod(..))
-import Hydra.Data.Party (Party(..), partyFromVerificationKeyBytes)
-import Test.QuickCheck (Arbitrary(..), vector)
+import Hydra.Data.ContestationPeriod (ContestationPeriod (..))
+import Hydra.Data.Party (Party (..), partyFromVerificationKeyBytes)
+import Test.QuickCheck (Arbitrary (..), vector)
 
 instance Arbitrary ContestationPeriod where
   arbitrary = fromInteger <$> arbitrary
