+++ hydra-tx/src/Hydra/Chain/ChainState.hs
@@ -5,12 +5,11 @@ module Hydra.Chain.ChainState where
 import Hydra.Prelude
 
 import Hydra.Tx (IsTx (..))
-import Test.Hydra.Prelude (Arbitrary)
 
 -- | A generic description for a chain slot all implementations need to use.
 newtype ChainSlot = ChainSlot Natural
   deriving stock (Ord, Eq, Show, Generic)
-  deriving newtype (Num, ToJSON, FromJSON, Arbitrary)
+  deriving newtype (Num, ToJSON, FromJSON)
 
 -- | Types that can be used on-chain by the Hydra protocol.
 -- XXX: Find a better name for this. Maybe IsChainTx or IsL1Tx?
@@ -20,7 +19,6 @@ class
   , Show (ChainStateType tx)
   , FromJSON (ChainStateType tx)
   , ToJSON (ChainStateType tx)
-  , Arbitrary (ChainStateType tx)
   ) =>
   IsChainState tx
   where
