+++ hydra-tx/src/Hydra/Tx/Increment.hs
@@ -5,7 +5,6 @@ import Hydra.Prelude
 
 import Cardano.Api.UTxO qualified as UTxO
 import Data.List qualified as List
-import Hydra.Cardano.Api.Gen ()
 import Hydra.Contract.Deposit qualified as Deposit
 import Hydra.Contract.Head qualified as Head
 import Hydra.Contract.HeadState qualified as Head
