+++ hydra-tx/src/Hydra/Tx/IsTx.hs
@@ -24,7 +24,6 @@ import Hydra.Cardano.Api.Tx qualified as Api
 import Hydra.Cardano.Api.UTxO qualified as Api
 import Hydra.Contract.Util qualified as Util
 import PlutusLedgerApi.V3 (fromBuiltin)
-import Test.QuickCheck (Arbitrary (..))
 
 -- | Types of transactions that can be used by the Head protocol. The associated
 -- types and methods of this type class represent the whole interface of what
@@ -96,16 +95,6 @@ class
 
   -- | Return the left-hand side without the right-hand side.
   withoutUTxO :: UTxOType tx -> UTxOType tx -> UTxOType tx
-
--- * Constraint synonyms
-
-type ArbitraryIsTx tx =
-  ( IsTx tx
-  , Arbitrary tx
-  , Arbitrary (UTxOType tx)
-  , Arbitrary (TxIdType tx)
-  , Arbitrary (TxOutType tx)
-  )
 
 -- * Cardano Tx
