+++ hydra-tx/test/Hydra/Tx/Contract/Abort.hs
@@ -5,9 +5,9 @@
 module Hydra.Tx.Contract.Abort where
 
 import Hydra.Cardano.Api
+import Hydra.Plutus.Gen ()
 import Hydra.Prelude
 import Test.Hydra.Prelude
-import Hydra.Plutus.Gen ()
 
 import Cardano.Api.UTxO qualified as UTxO
 import Data.List qualified as List
