+++ hydra-tx/test/Hydra/Tx/Contract/CollectCom.hs
@@ -4,9 +4,9 @@
 module Hydra.Tx.Contract.CollectCom where
 
 import Hydra.Cardano.Api
+import Hydra.Plutus.Gen ()
 import Hydra.Prelude hiding (label, toList)
 import Test.Hydra.Prelude
-import Hydra.Plutus.Gen ()
 
 import Cardano.Api.UTxO qualified as UTxO
 import Data.List qualified as List
