+++ hydra-tx/test/Hydra/Tx/Contract/Increment.hs
@@ -3,9 +3,9 @@
 module Hydra.Tx.Contract.Increment where
 
 import Hydra.Cardano.Api
+import Hydra.Plutus.Gen ()
 import Hydra.Prelude hiding (label)
 import Test.Hydra.Prelude
-import Hydra.Plutus.Gen ()
 
 import Cardano.Api.UTxO qualified as UTxO
 import Data.Maybe (fromJust)
