+++ hydra-tx/test/Hydra/Tx/Contract/Init.hs
@@ -15,11 +15,11 @@ import Hydra.Contract.HeadState (State (..))
 import Hydra.Contract.HeadTokensError (HeadTokensError (..))
 import Hydra.Tx.HeadParameters (HeadParameters (..))
 import Hydra.Tx.Init (initTx)
-import Hydra.Tx.OnChainId (OnChainId, genOnChainId)
+import Hydra.Tx.OnChainId (OnChainId)
 import Hydra.Tx.Party (Party)
 import PlutusLedgerApi.Test.Examples qualified as Plutus
 import Test.Hydra.Tx.Fixture (testNetworkId, testPolicyId, testSeedInput)
-import Test.Hydra.Tx.Gen (genForParty, genOneUTxOFor, genValue)
+import Test.Hydra.Tx.Gen (genForParty, genOnChainId, genOneUTxOFor, genValue)
 import Test.Hydra.Tx.Mutation (
   Mutation (..),
   SomeMutation (..),
@@ -28,7 +28,7 @@ import Test.Hydra.Tx.Mutation (
   modifyInlineDatum,
   replaceHeadId,
  )
-import Test.QuickCheck (Gen, choose, elements, oneof, suchThat, vectorOf)
+import Test.QuickCheck (choose, elements, oneof, suchThat, vectorOf)
 import Prelude qualified
 
 --
